gis_project/requirement.txt
asgiref==3.8.1
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==42.0.5
defusedxml==0.8.0rc2
dj-database-url==2.1.0
Django==3.2.10
django-allauth==0.63.3
django-cors-headers==4.3.1
django-crispy-forms==2.1
django-extensions==3.2.3
django-heroku==0.3.1
django-ranged-response==0.2.0
django-recaptcha==4.0.0
django-rest-auth==0.9.5
django-simple-captcha==0.6.0
django-six==1.0.5
django-tinymce==4.0.0
djangorestframework==3.15.1
djangorestframework-gis==1.0
fontawesome-free==5.15.4
fontawesomefree==6.5.1
gunicorn==21.2.0
idna==3.6
Jinja2==3.1.4
MarkupSafe==2.1.5
mysqlclient==2.2.4
numpy==1.26.4
oauthlib==3.2.2
packaging==24.0
pandas==2.2.2
pillow==10.2.0
psycopg2==2.9.9
psycopg2-binary==2.9.9
pycparser==2.22
PyJWT==2.8.0
python-dateutil==2.9.0.post0
python-linkedin==4.1
python3-openid==3.2.0
pytz==2024.1
requests==2.31.0
requests-oauthlib==2.0.0
six==1.16.0
social-auth-app-django==5.4.0
social-auth-core==4.5.3
sqlparse==0.4.4
typing_extensions==4.10.0
tzdata==2024.1
urllib3==2.2.1
whitenoise==6.6.0
